20090331 SimulationModal toppanel alfa version.

+++ Dynamo-HIA_UI/src/nl/rivm/emi/dynamo/data/types/atomic/base/AbstractRangedInteger.java
@@ -140,7 +140,7 @@ abstract public class AbstractRangedInteger extends
 
 		public Object getToType() {
 			// log.debug(debugString + " getToType() entered.");
-			return (Object) Float.class;
+			return (Object) Integer.class;
 		}
 	}
